+++ app/forms/transfer/edit_project_form.rb
@@ -9,6 +9,7 @@ class Transfer::EditProjectForm
   attribute :outgoing_trust_sharepoint_link
   attribute :incoming_trust_sharepoint_link
   attribute :outgoing_trust_ukprn
+  attribute :incoming_trust_ukprn
 
   validates :establishment_sharepoint_link, presence: true, sharepoint_url: true
   validates :incoming_trust_sharepoint_link, presence: true, sharepoint_url: true
@@ -17,13 +18,19 @@ class Transfer::EditProjectForm
   validates :outgoing_trust_ukprn, presence: true, ukprn: true
   validates :outgoing_trust_ukprn, trust_exists: true, if: -> { outgoing_trust_ukprn.present? }
 
+  validates :incoming_trust_ukprn, presence: true, ukprn: true
+  validates :incoming_trust_ukprn, trust_exists: true, if: -> { incoming_trust_ukprn.present? }
+
+  validates_with OutgoingIncomingTrustsUkprnValidator
+
   def self.new_from_project(project)
     new(
       project: project,
       establishment_sharepoint_link: project.establishment_sharepoint_link,
       incoming_trust_sharepoint_link: project.incoming_trust_sharepoint_link,
       outgoing_trust_sharepoint_link: project.outgoing_trust_sharepoint_link,
-      outgoing_trust_ukprn: project.outgoing_trust_ukprn
+      outgoing_trust_ukprn: project.outgoing_trust_ukprn,
+      incoming_trust_ukprn: project.incoming_trust_ukprn
     )
   end
 
@@ -36,7 +43,8 @@ class Transfer::EditProjectForm
       establishment_sharepoint_link: establishment_sharepoint_link,
       incoming_trust_sharepoint_link: incoming_trust_sharepoint_link,
       outgoing_trust_sharepoint_link: outgoing_trust_sharepoint_link,
-      outgoing_trust_ukprn: outgoing_trust_ukprn
+      outgoing_trust_ukprn: outgoing_trust_ukprn,
+      incoming_trust_ukprn: incoming_trust_ukprn
     )
     if valid?
       project.save